\name{BagseCpp-package}
\alias{BagseCpp-package}
\alias{BagseCpp}
\docType{package}
\title{
\packageTitle{BagseCpp}
}
\description{
\packageDescription{BagseCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BagseCpp}
\packageIndices{BagseCpp}
~~ Package for enrichment analysis ~~
}
\author{
\packageAuthor{BagseCpp}

Maintainer: \packageMaintainer{BagseCpp}
}
\references{
~~ Literature or other references for background information ~~
}
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
##~~ simple examples of the most important functions ~~
}
