% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log10BF}
\alias{compute_log10BF}
\title{An Rcpp Armadillo function to find log10 Bayes factors for an individual gene corresponding to a grid of a mixture normal}
\usage{
compute_log10BF(bhat, sdhat, phi_vec)
}
\arguments{
\item{bhat}{A double that represents an estimated effect size of a single gene}

\item{sdhat}{A double that represents the standard error of the estimated effect size of a single gene}

\item{phi_vec}{A k length column vector type that represents the standard deviations of the mixture of normal distributions distribution of the true effect size of an associated gene}
}
\value{
A k length column vector of log10 Bayes Factors, representing the likelihood of the gene coming from that that component of the mixture normal
}
\description{
An Rcpp Armadillo function to find log10 Bayes factors for an individual gene corresponding to a grid of a mixture normal
}
