% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computegridwts}
\alias{computegridwts}
\title{Rcpp Armadillo function to compute a vector containing posterior weights for the mixture normal distribution of an associated gene's effect size}
\usage{
computegridwts(log10_BF_vec, pi_vec, pi0, log10_NC)
}
\arguments{
\item{log10_BF_vec}{A k length vector of Bayes Factors representing likelihoods of the gene coming from each of the p mixture normal components}

\item{pi_vec}{A k length vector of prior probabilities of the gene coming from each of the p mixture normal components}

\item{pi0}{A double type prior probability of the gene being unassociated}

\item{log10_NC}{A double type number, which is log10 of a normalizing constant, for the logsum of the prior probabilities and the Bayes Factors}
}
\value{
A k+1 length vector of posterior probabilities of the gene coming from each component of the mixture normal. The first entry is the posterior probability of the gene being unassociated
}
\description{
Rcpp Armadillo function to compute a vector containing posterior weights for the mixture normal distribution of an associated gene's effect size
}
