% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{torus_cpp}
\alias{torus_cpp}
\title{An Rcpp Armadillo function to calculate enrichment parameters, local FDR for all genes, and estimates for the weights of the components for the mixture normal for associated genes}
\usage{
torus_cpp(betahat, sebetahat, annotation, tol = 0.1)
}
\arguments{
\item{betahat}{An n length column vector containing effect size estimates for all n genes}

\item{sebetahat}{An n length column vector containing the standard errors of the effect size estimates for all n genes}

\item{annotation}{An n length integer vector of 0s and 1s, with a 1 corresponding to a gene in the gene set of interest}

\item{tol}{A double type number corresponding to the threshold for when the EM algorithm should converge. Default is 1e-1}
}
\value{
A list containing the following variables:
* enrichment_est     A 2 length column vector containing the enrichment parameter estimates
* effect_est         An column vector containing estimates for the weights of the mixture normal of the effect size for associated genes 
* EM                 The raw output from the EM algorithm within this function
* ldfr               An n length column vector containing the local FDR for all genes, using enrichment information
}
\description{
An Rcpp Armadillo function to calculate enrichment parameters, local FDR for all genes, and estimates for the weights of the components for the mixture normal for associated genes
}
