% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{torus_pool_em}
\alias{torus_pool_em}
\title{An Rcpp Armadillo function to compute an iteration of the key EM algorithm}
\usage{
torus_pool_em(p, BF_matrix, annot_vec)
}
\arguments{
\item{p}{A k+2 length column vector, where k is the components in the mixture normal, that contains all the parameters to be estimated, including the enrichment parameters and the mixture normal weights}

\item{BF_matrix}{A (n x k) matrix of Bayes Factors, which is the k Bayes Factors representing each mixture normal component for all n genes}

\item{annot_vec}{A n length integer vector of 0s and 1s where a 1 represents that the gene is in the gene set of interest}
}
\value{
A k+2 length column vector with the new parameter estimates, with the first two being the enrichment parameters and the rest being the mixture normal weight estimates
}
\description{
An Rcpp Armadillo function to compute an iteration of the key EM algorithm
}
