% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{torus_pool_loglik}
\alias{torus_pool_loglik}
\title{An Rcpp Armadillo function to find the log likelihood using the Bayesian model, with a given set of parameters}
\usage{
torus_pool_loglik(p, BF_matrix, annot_vec)
}
\arguments{
\item{p}{A k+2 length column vector, where k is the components in the mixture normal, that contains all the parameters, including the enrichment parameters and the mixture normal weights}

\item{BF_matrix}{A (n x k) matrix of Bayes Factors, which is the k Bayes Factors representing each mixture normal component for all n genes}

\item{annot_vec}{A n length integer vector of 0s and 1s where a 1 represents that the gene is in the gene set of interest}
}
\value{
A double type representing the log likelihood with this set of parameters
}
\description{
An Rcpp Armadillo function to find the log likelihood using the Bayesian model, with a given set of parameters
}
