% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedlog10BF}
\alias{weightedlog10BF}
\title{An Rcpp Armadillo function that computes a weighted sum of log10 Bayes Factors}
\usage{
weightedlog10BF(log10_BFv, wv)
}
\arguments{
\item{log10_BFv}{A k length column vector with log10 Bayes Factors corresponding to components of the mixture normal for the effect size of a gene}

\item{wv}{A k length column vector of weights corresponding to the components of the mixture normal for the effect size of a gene}
}
\value{
The weighted sum of the log10 Bayes Factors of the components of the mixture normal for the effect size of a gene
}
\description{
An Rcpp Armadillo function that computes a weighted sum of log10 Bayes Factors
}
